INTEGER型をINT型に修正

+++ db/db-challenge2/chatwork.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    user_id INTEGER(11) PRIMARY KEY AUTO_INCREMENT,
+    user_id INT(11) PRIMARY KEY AUTO_INCREMENT,
     email_address VARCHAR(100) NOT NULL,
     password VARCHAR(100) NOT NULL,
     nickname VARCHAR(100) NOT NULL,
@@ -12,7 +12,7 @@ CREATE TABLE users (
 );
 
 CREATE TABLE chatrooms (
-    chat_room_id INTEGER(11) PRIMARY KEY AUTO_INCREMENT,
+    chat_room_id INT(11) PRIMARY KEY AUTO_INCREMENT,
     chat_name VARCHAR(100) NOT NULL,
     chat_description VARCHAR(1000),
     is_file_sendable TINYINT(1) DEFAULT 1 NOT NULL,
@@ -25,34 +25,34 @@ CREATE TABLE chatrooms (
 );
 
 CREATE TABLE chatposts (
-    chat_post_id INTEGER(11) PRIMARY KEY AUTO_INCREMENT,
-    chat_room_id INTEGER(11) NOT NULL REFERENCES chatrooms(chat_room_id),
+    chat_post_id INT(11) PRIMARY KEY AUTO_INCREMENT,
+    chat_room_id INT(11) NOT NULL REFERENCES chatrooms(chat_room_id),
     message VARCHAR(1000) NOT NULL,
     attached_filename VARCHAR(100),
     is_deleted TINYINT(1) NOT NULL DEFAULT 0,
     created_at DATETIME NOT NULL,
-    owner_user_id INTEGER(11) NOT NULL REFERENCES users(user_id),
+    owner_user_id INT(11) NOT NULL REFERENCES users(user_id),
     updated_at DATETIME NOT NULL,
-    update_user_id INTEGER(11) NOT NULL REFERENCES users(user_id)
+    update_user_id INT(11) NOT NULL REFERENCES users(user_id)
 );
 
 CREATE TABLE chatmembers (
-    chat_room_id INTEGER(11) NOT NULL REFERENCES chatrooms(chat_room_id),
+    chat_room_id INT(11) NOT NULL REFERENCES chatrooms(chat_room_id),
     member_user_id VARCHAR(255) NOT NULL REFERENCES users(user_id),
     PRIMARY KEY(chat_room_id, member_user_id),
     joined_at DATETIME NOT NULL
 );
 
 CREATE TABLE tasks (
-    task_id INTEGER(11) PRIMARY KEY AUTO_INCREMENT,
+    task_id INT(11) PRIMARY KEY AUTO_INCREMENT,
     task_description VARCHAR(1000) NOT NULL,
-    task_assigned_id INTEGER(11) NOT NULL REFERENCES users(user_id),
+    task_assigned_id INT(11) NOT NULL REFERENCES users(user_id),
     due_date DATETIME,
-    chat_room_id INTEGER(11) NOT NULL REFERENCES chatrooms(chat_room_id),
+    chat_room_id INT(11) NOT NULL REFERENCES chatrooms(chat_room_id),
     is_completed TINYINT(1) NOT NULL DEFAULT 0,
     is_deleted TINYINT(1) NOT NULL DEFAULT 0,
     created_at DATETIME NOT NULL,
-    owner_user_id INTEGER(11) NOT NULL REFERENCES users(user_id),
+    owner_user_id INT(11) NOT NULL REFERENCES users(user_id),
     updated_at DATETIME NOT NULL,
-    update_user_id INTEGER(11) NOT NULL REFERENCES users(user_id)
+    update_user_id INT(11) NOT NULL REFERENCES users(user_id)
 );
